refactor: 디렉토리 구조 및 파일명 변경
