--- src/js/models/index.js
@@ -1,3 +1,0 @@
-// 레이싱 게임 영역
-export { default as RacingCarInstance } from './RacingCar.js';
-export { default as RacingGameModel } from './RacingGame.js';
